+++ x/upgrade/keeper/keeper.go
@@ -29,7 +29,6 @@ type Keeper struct {
 	storeKey           sdk.StoreKey
 	cdc                codec.BinaryMarshaler
 	upgradeHandlers    map[string]types.UpgradeHandler
-	versionMap         module.VersionMap
 }
 
 // NewKeeper constructs an upgrade Keeper
@@ -50,20 +49,8 @@ func (k Keeper) SetUpgradeHandler(name string, upgradeHandler types.UpgradeHandl
 	k.upgradeHandlers[name] = upgradeHandler
 }
 
-// SetInitialVersionMap sets an initial version map on the keeper.
-// This must be set from app.go after module Manager is initialized.
-func (k *Keeper) SetInitialVersionMap(vm module.VersionMap) {
-	k.versionMap = vm
-}
-
-// SetCurrentConsensusVersions saves the current version map
-// to state from InitGenesis
-func (k Keeper) SetCurrentConsensusVersions(ctx sdk.Context) {
-	k.setConsensusVersions(ctx, k.versionMap)
-}
-
-// setConsensusVersions saves a given version map to state
-func (k Keeper) setConsensusVersions(ctx sdk.Context, vm module.VersionMap) {
+// SetConsensusVersions saves a given version map to state
+func (k Keeper) SetConsensusVersions(ctx sdk.Context, vm module.VersionMap) {
 	if len(vm) > 0 {
 		store := ctx.KVStore(k.storeKey)
 		versionStore := prefix.NewStore(store, []byte{types.VersionMapByte})
@@ -239,7 +226,7 @@ func (k Keeper) ApplyUpgrade(ctx sdk.Context, plan types.Plan) {
 		panic(err)
 	}
 
-	k.setConsensusVersions(ctx, updatedVM)
+	k.SetConsensusVersions(ctx, updatedVM)
 
 	// Must clear IBC state after upgrade is applied as it is stored separately from the upgrade plan.
 	// This will prevent resubmission of upgrade msg after upgrade is already completed.